Created new file for ActionState
